This commit address a minor bug in the ideal HVAC model. When simulating a system which uses a GSHP for heating and a ASHP for cooling, the subroutine GSHPSIM in cetc/gshpsim.F returns a 0 COP during a cooling period while calculating the daily values for the GSHP. This COP=0 is then used as a denominator when determining cooling load dQ, causing a return of inf or NaN. WORK IN PROGRESS: Further testing required to ensure ASHP and GSHP are interoperable and are returning reasonable values.
Also, calls to subrountine FREEZING in gshpsim.F have been commented out. Based on inspection
of the source code, this subroutine returns the same values it is sent. WORK IN PROGRESS:
the logic of this subroutine needs to be re-assessed to account for freezing fluid
corrections.

Summary of Changes:
- Added if statements to determine if the COP returned by the function daily_GSHP_COP
  is greater than 0. If it isn't, dQ=0.

Testing summary:
- Compiles on Cygwin/W7