## Initial commit of circleci and tester tests
This commit adds calls to tester in the config.yml file. Whenever a commit
is made to this repo, circleci is spun up, and ESP-r is compiled. Non-zero
exit codes were added to the install script to let circleci know when the
script has exited with an error. Tester is run via the RunTester.pl script
which pulls down a historical archive from Google drive, runs the compiled
ESP-r against the archive, and returns an exit code of zero if no
differences are found.