Enhancement to pass surface temps through co-simulator
This commit adds functionality to the ESP-r/TRNSYS co-simulator. Specified
internal zone surfaces may now send their temperature data to TRNSYS at
each timestep. This hack uses a similar approach introduced on
Sebastien_Brideau branch @r9363 with a key differences; an extra,
artificial zone is not required.

Instead, the user specifies the number of additional surface temperatures
to be read into TRNSYS as well as their zone and surface indexes at the
bottom of the plant network file.

Additionally the may user can use the explicit plant network menu in prj
to define how many and which internal surface temperatures in the model to
send to TRNSYS via the harmonizer.

Testing summary:
- Successfully compiled X11 version on Windows XP/Cygwin (GCC 4.5.3).
- Successfully compiled GTK version on Windows XP.
- Successfully compiled and co-simulated with bps.dll
- Compared surface temperatures received by TRNSYS and the h3k output
- Opened a pre-configured .pln file with co-simulation surface linkage
  defined.
- Created a .pln file with no co-simulation surface linkages defined
  using the plant network menu items.
- Created a .pln file with co-simulation surface linkages defined
  using the plant network menu items.