Bug fix for TDF facility
When running annual simulations with an annual TDF, the subroutine RCTDFB in tdrecbps.F was not indexing
the first hour of January first properly. For the first hour of January, the data from the first hour
of December 31st was being read. I change to logic in tdrecbps.F was implemented to address this.

The variable iknowrec in RCTDFB was also statically allocated to hold up to 87600 elements. This corresponded
to a one year simulation period at a ten minute timestep, with no warm-up. iknowrec was moved to a new module
tdrecs where it can be dynamically allocated based on total number of simulation timesteps. Initialization of
this array was moved from input.F to this module

WORK IN PROGRESS: Checks and warnings need to be implemented for cases where the simulation period is outside
the TDF period. Testing of the interpolation is also required. Currently, all testing has been done for the
CASUALT item type only.