Forced same timesteps per hour between simulation and TDF file
The TDF facility has some coded capability for interpolating. This commit forces that the TDF file and simulation have the same timstep. If they do not, a warning will be issued and the simulation will halt.

TODO: Determine if it's necessary to enable interpolation. For now, disable it. It is now up to the user to pre-process the TDF to align with desired simulation tstep.