Updated heat gain calculation for explicit plant network fan
Prior to this change, the total absorbed electrical energy determined each timestep in CMP03C was also set as heat gain to the node. Per equation 5.10a in Jan Hensen's thesis, the heat gain 'Q' due to inefficiency should be Q=E*(1-eff) where 'E' is the electrical power absorbed and 'eff' is the efficiency of the fan. This commit adds this change.