Updated archived tester results
The last commit made changes to how heat gain from fan inefficiency is calculated for the explicit plant fan component. This caused changes in results for tester models that used the fan. This commit points the tester script used by circleCI to an updated archive of tester results that was generated using the new fan heat gain calculation.