Updated master/slave zone control
It was found in the source code that only BCL01 and BCL02 can act as a master control to BCL21. No warnings are given, and the zones under slave free-float. This commit enables BCL10 (ON/OFF) to be used with BCL21. A warning is thrown once per offending control loop if the master loop is not a compatible type, and the slave zones free float.